Set-DefaultGateway Updates
Tweaked output in Set-DefaultGateway to allow for more prescriptive errors if a /24 isn't being used on the netmask vs if a fourth octet isn't provided (assumes a gateway wasn't passed in).

Added Pester tests for Set-DefaultGateway -- should be working!

Added not-yet-working Pester test for Confirm-BackingNetwork.... logic isn't quite there yet.